Module2/requirements_face.txt
gradio==4.19.2
huggingface-hub==0.20.3
ultralytics==8.3.159
Pillow==11.2.1
numpy==1.26.4
opencv-python==4.11.0.86 